+++ src/main/java/up/edu/microservicios/model/Paciente.java
@@ -6,12 +6,12 @@ public class Paciente {
     private Integer id;
     private String  nombre;
     private String apellido;
-    private Integer numeroContacto;
+    private Long  numeroContacto;
     private LocalDate fechaIngreso;
     private Domicilio domicilio;
     private String email;
 
-    public Paciente(Integer id, String nombre, String apellido, Integer numeroContacto, LocalDate fechaIngreso, Domicilio domicilio, String email) {
+    public Paciente(Integer id, String nombre, String apellido, Long  numeroContacto, LocalDate fechaIngreso, Domicilio domicilio, String email) {
         this.nombre = nombre;
         this.apellido = apellido;
         this.numeroContacto = numeroContacto;
@@ -21,7 +21,7 @@ public class Paciente {
         this.id = id;
     }
 
-    public Paciente(String nombre, String apellido, Integer numeroContacto, LocalDate fechaIngreso, Domicilio domicilio, String email) {
+    public Paciente(String nombre, String apellido, Long  numeroContacto, LocalDate fechaIngreso, Domicilio domicilio, String email) {
         this.nombre = nombre;
         this.apellido = apellido;
         this.numeroContacto = numeroContacto;
@@ -56,11 +56,11 @@ public class Paciente {
         this.apellido = apellido;
     }
 
-    public Integer getNumeroContacto() {
+    public Long  getNumeroContacto() {
         return numeroContacto;
     }
 
-    public void setNumeroContacto(Integer numeroContacto) {
+    public void setNumeroContacto(Long  numeroContacto) {
         this.numeroContacto = numeroContacto;
     }
